upload in mobile app

+++ src/app/component/protect-component/customers/component/customer/accounts/assets/mutual-fund/tempservice.service.ts
@@ -588,7 +588,7 @@ export class TempserviceService {
 
   mutualFundRoundAndFormat(data, noOfPlaces: number = 0) {
     if (data) {
-      data = parseInt(data)
+      data = parseFloat(data)
       if (isNaN(data)) {
         return data;
       } else {
